Include repaint in image load time measurement
There seems to be some 'trickery'
behind the scenes. When the image is displayed
for the first time there is noticeable delay. This
is not due to file load as this has been done in the
background. But when it is displayed again it shows
instantly. Seems that there is some form of caching
happening in the background.