Do not call image load directly from UI
Enables really fast cruising through directories with a lot of files.
Before if user holds A,D, Left or right UI would be blocked by huge
amount of image loads. Now only file index is changed if we hold
one of the buttons (which is super fast).
And when a user stops that image is loaded and shown.

ImagePanel.java
* when user changes file index do not invoke file load directly
* instead just change desire file index
* this change of file index is picked up in background thread
* when desired file is ready thread will notify UI
* when UI receives info it will display new image

ImageLoaderThread
* add ImagePanel reference
* detect change in file index and load appropriate file
* after loading notify ImagePanel that the image is ready

Main.java
* add test label to show current file index
* probably will become permanent as it is really useful