Restore cut off parts of the image on init
If image file has EXIF rotate data initially parts of it would be
cut off. Reason for this is that x,y coordinates as well as scaling
ratio was calculated using original width and height of the image.
Not that if the image is rotated width and height essentially change
places so the calculation needs to take this into account.